heatmap: refactor component structure - set test skeleton
